test: migrate internal-event-consumers to bdd
